pyproject.toml
[build-system]
requires = ["uv_build>=0.8.13"]
build-backend = "uv_build"

[project]
name = "pdf_parsers"
version = "0.2.0"
readme = "README.md"
description = "Parse PDFs for information extraction"
license = { file = "LICENSE" }
authors = [{ name = "Raphael LaFrance", email = "raphael.lafrance@gmail.com" }]
requires-python = ">=3.13"
dependencies = [
    "lmstudio>=1.5.0",
    "pillow>=12.0.0",
    # "Jinja2",
    # "pillow",
    # "regex",
    # "rich",
    "rich>=14.2.0",
    "tqdm>=4.67.1",
]

[dependency-groups]
dev = ["pre-commit>=4.3.0", "uv-build>=0.8.13"]

[tool.uv.build-backend]
module-root = "."
module-name = "parse"

[tool.uv]

[tool.ruff]

show-fixes = true

[tool.ruff.lint]

select = ["ALL"]

ignore = [
    "C901",    # is too complex
    "COM812",  # Trailing comma missing
    "D100",    # Missing docstring in public module
    "D101",    # Missing docstring in public class
    "D102",    # Missing docstring in public method
    "D103",    # Missing docstring in public function
    "D104",    # Missing docstring in public package
    "D105",    # Missing docstring in public method
    "D107",    # Missing docstring in public __init__
    "D203",    # 1 blank line required before class docstring
    "D212",    # Multi-line docstring summary should start at the first line
    "DTZ005",  # Pass a `datetime.timezone` object to the `tz` parameter
    "ERA001",  # Remove commented-out code
    "FIX002",  # Line contains TODO, consider resolving the issue
    "LOG015",  # Use own logger instead TODO
    "PLW2901", # Outer {outer_kind} variable {name} overwritten by inner {inner_kind} target
    "PT009",   # Use a regular "assert" instead of unittest-style {assertion}
    "RET504",  # Unnecessary assignment to {name} before return statement
    "RUF001",  # String contains ambiguous {}. Did you mean {}?
    "RUF003",  # Comment contains ambiguous character
    "SIM114",  # Combine if branches using logical or operator
    "SIM117",  # Use a single with statement with multiple contexts instead of nested with statements
    "SLF001",  # Private member accessed: {access}
    "T201",    # print found
    "TD002",   # Missing author in TODO
    "TD003",   # Missing issue link for this TODO
]
